+++ R/weight.R
@@ -99,11 +99,18 @@ weight2 <- function(x, weights) {
 
 #' @title Weighted statistics for variables
 #' @name wtd_sd
-#' @description Compute weighted standard deviation or standard error for a
-#'                variable or for all variables of a data frame.
+#' @description \code{wtd_sd()} and \code{wtd_se()} compute weighted standard
+#'   deviation or standard error for a variable or for all variables of a data
+#'   frame. \code{svy_md()} computes the median for a variable in a survey-design
+#'   (see \code{\link[survey]{svydesign}}).
 #'
-#' @param x (Numeric) vector or a data frame.
+#' @param x (Numeric) vector or a data frame. For \code{svy_md()}, the bare
+#'          (unquoted) variable name, or a character vector with the variable
+#'          name.
 #' @param weights Numeric vector of weights.
+#'
+#' @inheritParams svyglm.nb
+#'
 #' @return The weighted standard deviation or standard error of \code{x},
 #'           or for each variable if \code{x} is a data frame.
 #'
@@ -114,6 +121,22 @@ weight2 <- function(x, weights) {
 #' data(efc)
 #' wtd_sd(efc[, 1:3], runif(n = nrow(efc)))
 #' wtd_se(efc[, 1:3], runif(n = nrow(efc)))
+#'
+#'
+#' # median for variables from weighted survey designs
+#' library(survey)
+#' data(nhanes_sample)
+#'
+#' des <- svydesign(
+#'   id = ~SDMVPSU,
+#'   strat = ~SDMVSTRA,
+#'   weights = ~WTINT2YR,
+#'   nest = TRUE,
+#'   data = nhanes_sample
+#' )
+#'
+#' svy_md(total, des)
+#' svy_md("total", des)
 #'
 #' @export
 wtd_sd <- function(x, weights = NULL) {
@@ -163,7 +186,7 @@ wtd_se <- function(x, weights = NULL) {
 
 #' @export
 wtd_se.data.frame <- function(x, weights = NULL) {
-  se_result <- purrr::map_dbl(x, ~ sqrt(Hmisc::wtd.var(.x, weights = weights, na.rm = TRUE) / length(stats::na.omit(.x))))
+  se_result <- purrr::map_dbl(x, ~ wtd_se_helper(.x, weights = weights))
   names(se_result) <- colnames(x)
 
   se_result
@@ -171,7 +194,7 @@ wtd_se.data.frame <- function(x, weights = NULL) {
 
 #' @export
 wtd_se.matrix <- function(x, weights = NULL) {
-  se_result <- purrr::map_dbl(x, ~ sqrt(Hmisc::wtd.var(.x, weights = weights, na.rm = TRUE) / length(stats::na.omit(.x))))
+  se_result <- purrr::map_dbl(x, ~ wtd_se_helper(.x, weights = weights))
   names(se_result) <- colnames(x)
 
   se_result
@@ -179,5 +202,33 @@ wtd_se.matrix <- function(x, weights = NULL) {
 
 #' @export
 wtd_se.default <- function(x, weights = NULL) {
+  wtd_se_helper(x, weights)
+}
+
+wtd_se_helper <- function(x, weights) {
   sqrt(Hmisc::wtd.var(x, weights = weights, na.rm = TRUE) / length(stats::na.omit(x)))
+}
+
+
+#' @rdname wtd_sd
+#' @importFrom stats as.formula
+#' @export
+svy_md <- function(x, design) {
+  # check if pkg survey is available
+  if (!requireNamespace("survey", quietly = TRUE)) {
+    stop("Package `survey` needed to for this function to work. Please install it.", call. = FALSE)
+  }
+
+  # deparse
+  v <- stats::as.formula(paste("~", as.character(substitute(x))))
+
+  as.vector(
+    survey::svyquantile(
+      v,
+      design = design,
+      quantiles = 0.5,
+      ci = FALSE,
+      na.rm = TRUE
+    )
+  )
 }